python/candy-connector/requirements.txt
pyusb==1.0.2
python-can==3.1.1
